Refactored common code to base class for FhirClient and Requester.
