+++ src/Hl7.Fhir.Specification.Tests/Source/ResourceResolverTests.cs
@@ -18,7 +18,6 @@ using System.IO;
 using Hl7.Fhir.Serialization;
 using System.Linq;
 using Hl7.Fhir.Utility;
-using System.Net.Http;
 
 namespace Hl7.Fhir.Specification.Tests
 {
@@ -95,13 +94,13 @@ namespace Hl7.Fhir.Specification.Tests
 
             public TestFhirClient(Uri endpoint) : base(endpoint) { Status = 1; }
 
-            protected override void BeforeRequest(HttpRequestMessage rawRequest, byte[] body)
+            protected override void BeforeRequest(HttpWebRequest rawRequest, byte[] body)
             {
                 Status = 2;
                 base.BeforeRequest(rawRequest, body);
             }
 
-            protected override void AfterResponse(HttpResponseMessage webResponse, byte[] body)
+            protected override void AfterResponse(HttpWebResponse webResponse, byte[] body)
             {
                 Status = 3;
                 base.AfterResponse(webResponse, body);
